chore: refactored on vscode realted code
