+++ scripts/dev.ts
@@ -3,7 +3,7 @@ import { remove } from 'fs-extra'
 
 async function dev() {
   await remove('./dist')
-  execSync('tsup src/index.ts --format cjs --external vscode --no-shims --watch', { stdio: 'inherit' })
+  execSync('tsup src/extension.ts --format cjs --external vscode --no-shims --watch', { stdio: 'inherit' })
 }
 
 dev()
