+++ src/actions.ts
@@ -2,7 +2,7 @@ import { Uri } from 'vscode'
 import { pForever } from './extrn'
 import { Log, closeAllOpenedFiles, getWorkspaceFolders, navigateFile, openTextDocument, prompt, showTextDocumentNonPreview } from './utils'
 import { getFiles, getGitIgnoreItems } from './utils/fs'
-import { config } from './config'
+import { config } from './utils/vscode/config'
 
 export async function start() {
   Log.info(`Configs: ${JSON.stringify(config)}`)
