+++ src/utils/index.ts
@@ -1,5 +1,5 @@
-export * from './log'
-export * from './types'
 export * from './vscode'
+export * from './types'
+export * from './obj'
 
 // export * from './fs' // TODO: Need to work - can't be exported due to unit testing issue
