+++ src/utils/vscode/config.ts
@@ -1,8 +1,8 @@
 import { computed, reactive, ref } from '@vue/reactivity'
 import { workspace } from 'vscode'
-import { contributes } from '../package.json'
-import { EXT_NAMESPACE } from './meta'
-import { getProperty } from './utils/obj'
+import { contributes } from '../../../package.json'
+import { EXT_NAMESPACE } from '../../meta'
+import { getProperty } from '../obj'
 
 const _configState = ref(0)
