+++ src/utils/vscode/editor.ts
@@ -1,0 +1,5 @@
+import { window } from 'vscode'
+
+export function getActiveTextEditor() {
+  return window.activeTextEditor
+}
