+++ src/utils/vscode/log.ts
@@ -1,6 +1,6 @@
 import type { OutputChannel } from 'vscode'
 import { window } from 'vscode'
-import { EXT_NAME } from '../meta'
+import { EXT_NAME } from '../../meta'
 
 export class Log {
   private static _channel: OutputChannel
